added more unit tests
