+++ tinyscpi/scpi_parser.py
@@ -1,6 +1,7 @@
 import re
 
-import dictionaries.scpi_valid_dict as scpi_valid_dict
+from tinyscpi.dictionaries import scpi_valid_dict
+
 class SCPI_Parser:
     # what if parser drops every single lower case insturctions?
