+++ tinyscpi/tests/__init__.py
@@ -1,0 +1,2 @@
+# INTENTIONALLY LEFT BLANK
+# since unittest requires init.py within tests folder
